+++ src/components/navigation-bar/navigation-bar.stories.js
@@ -1,47 +1,36 @@
 import React from 'react';
-import { storiesOf } from '@storybook/react';
-import { text, select } from '@storybook/addon-knobs';
-import { dlsThemeSelector, classicThemeSelector } from '../../../.storybook/theme-selectors';
-import NavigationBar from './navigation-bar';
-import notes from './notes.md';
-import './navigation-bar.stories.scss';
-import getDocGenInfo from '../../utils/helpers/docgen-info';
+import { withKnobs, text } from '@storybook/addon-knobs';
+import NavigationBar from './navigation-bar.component';
 
-NavigationBar.__docgenInfo = getDocGenInfo(
-  require('./docgenInfo.json'),
-  /navigation-bar\.js(?!spec)/
-);
-
-function makeStory(name, themeSelector, disableChromatic = false) {
-  const component = () => {
-    const children = text('children', '');
-    const as = select(
-      'as',
-      ['primary', 'secondary', 'transparent'],
-      NavigationBar.defaultProps.as
-    );
-
-    return (
-      <NavigationBar
-        as={ as }
-      >
-        { children }
-      </NavigationBar>
-    );
-  };
-
-  const metadata = {
-    themeSelector,
-    info: { text: <p>Renders a full width application bar.</p> },
-    notes: { markdown: notes },
+export default {
+  title: 'Test/NavigationBar',
+  component: NavigationBar,
+  decorators: [withKnobs],
+  parameters: {
+    info: {
+      disable: true
+    },
+    knobs: { escapeHTML: false },
     chromatic: {
-      disable: disableChromatic
+      disable: true
     }
-  };
+  }
+};
 
-  return [name, component, metadata];
-}
+export const Basic = () => {
+  const children = text('children', 'Example content');
 
-storiesOf('Navigation Bar', module)
-  .add(...makeStory('default', dlsThemeSelector))
-  .add(...makeStory('classic', classicThemeSelector, true));
+  return (
+    <NavigationBar>
+      {children}
+    </NavigationBar>
+  );
+};
+
+Basic.story = {
+  parameters: {
+    chromatic: {
+      disable: true
+    }
+  }
+};